backend/requirements.txt
# Web 框架
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# 图像处理
Pillow==11.2.1
numpy==2.2.6

# 深度学习（CPU 版本）
torch==2.10.0
torchvision==0.21.0

# HuggingFace 生态
transformers==5.1.0
huggingface_hub==1.4.1
accelerate==1.12.0
safetensors==0.7.0

# 图像预处理
opencv-python-headless==4.11.0.86
